+++ NEWS.md
@@ -4,7 +4,11 @@ will continue until then.  These will be especially noticable around
 the inclusion of IV NCA parameters and additional specifications of
 the dosing including dose amount and route.
 
-# PKNCA 0.9.1.9000 (not released)
+# PKNCA 0.9.1.9000+ (not released)
+
+* New feature: the `time_calc()` function will help convert time
+  values to be relative to events (such as calculating time after and
+  before doses)
 
 # PKNCA 0.9.1